backend/pyproject.toml
[project]
name = "ai020-backend"
version = "0.1.0"
description = "AI020 Municipal Meeting Summarizer — backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "pypdf>=4.0.0",
    "pdfplumber>=0.11.0",
    "scikit-learn>=1.5.0",
    "jinja2>=3.1.0",
    "premailer>=3.10.0",
    "aiofiles>=24.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "types-passlib",
    "types-python-jose",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []  # app, not a library — no Python packages to install
